Add controller "Cart" -> CartsController Add script "Add to Cart" Add views for Cart Add routes for Cart
Update another views
Update controller OrdersController
Update controller ProductsController
Update controller ProductPhotosController
Update model Order
Update migrations for products, orders and reference order-products
Update seeder OrderProductSeeder

Update README.md

+++ database/migrations/2021_02_23_083914_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             Schema::create('orders', function (Blueprint $table) {
                 $table->id();
 
-                $table->integer('user_own_id')->unsigned()->nullable();
+                $table->bigInteger('user_own_id')->unsigned()->nullable();
 
                 $table->string('uuid', 36)->unique()->nullable(false);
 //                $table->string('slug', 255)->nullable(false);
